+++ packages/backend/src/database/entities/dashboards.ts
@@ -12,8 +12,6 @@ export const DashboardTilesTableName = 'dashboard_tiles';
 export const DashboardTileTypesTableName = 'dashboard_tile_types';
 export const DashboardTileChartTableName = 'dashboard_tile_charts';
 export const DashboardTileSqlChartTableName = 'dashboard_tile_sql_charts';
-export const DashboardTileSemanticViewerChartTableName =
-    'dashboard_tile_semantic_viewer_charts';
 export const DashboardTileMarkdownsTableName = 'dashboard_tile_markdowns';
 export const DashboardTileLoomsTableName = 'dashboard_tile_looms';
 export const DashboardTabsTableName = 'dashboard_tabs';